feat: added biometric lock for password visibility and updatability

+++ app/src/main/java/com/example/keyhive/components/EditPasswordFormComponent.kt
@@ -3,6 +3,7 @@ package com.example.keyhive.components
 
 import android.provider.CalendarContract.Colors
 import android.widget.Toast
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -13,12 +14,15 @@ import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material.icons.filled.Done
 import androidx.compose.material.icons.filled.Update
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Checkbox
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.FilledTonalButton
+import androidx.compose.material3.FilterChip
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Switch
 import androidx.compose.material3.Text
@@ -35,7 +39,9 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.example.keyhive.model.Password
+import com.example.keyhive.routes.Routes
 import com.example.keyhive.ui.theme.RedA100
+import com.example.keyhive.ui.theme.RedA700
 import com.example.keyhive.utils.CryptoUtils
 import com.example.keyhive.viewmodel.PasswordViewModel
 import java.util.Date
@@ -50,6 +56,7 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
     val favoriteState = rememberSaveable() {
         mutableStateOf(password.isFavorite)
     }
+
     val keyboardController = LocalSoftwareKeyboardController.current
     val validateUserName = remember(userNameState.value){
         userNameState.value.trim().isNotEmpty()
@@ -64,10 +71,11 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
         typeState.value.trim().isNotEmpty()
     }
     val context = LocalContext.current
+    val showBiometricAuthDialog = remember {
+        mutableStateOf(false)
+    }
 
-    Column(modifier = modifier){
-
-        Row(modifier = modifier.padding(5.dp)){
+    Column(modifier = modifier.padding(16.dp), verticalArrangement = Arrangement.spacedBy(10.dp) ){
             CommonTextField(
                 valueState = userNameState,
                 placeholder = "Username*",
@@ -79,8 +87,7 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
                     keyboardController?.hide()
                 }
             )
-        }
-        Row(modifier = modifier.padding(5.dp)){
+
             PasswordTextField(
                 valueState = passwordState,
                 placeholder = "Password*",
@@ -92,8 +99,8 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
                     keyboardController?.hide()
                 }
             )
-        }
-        Row(modifier = modifier.padding(5.dp)){
+
+
             CommonTextField(
                 valueState = typeState,
                 placeholder = "Website/App*",
@@ -105,8 +112,8 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
                     keyboardController?.hide()
                 }
             )
-        }
-        Row(modifier = modifier.padding(5.dp)){
+
+
             CommonTextField(
                 valueState = descriptionState,
                 placeholder = "Description",
@@ -120,34 +127,35 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
                 maxLines = 3,
                 singleLine = false
             )
+        Column(modifier = modifier.padding(5.dp),
+            horizontalAlignment = Alignment.Start,
+            verticalArrangement = Arrangement.spacedBy(8.dp)){
+            Row(modifier = modifier.fillMaxWidth().padding(5.dp),
+                horizontalArrangement = Arrangement.spacedBy(10.dp)){
+                FilterChip(
+                    onClick = { enableBiometricAuthState.value = !enableBiometricAuthState.value },
+                    label = { Text("Lock with Biometric") },
+                    selected = enableBiometricAuthState.value,
+                    leadingIcon = if (enableBiometricAuthState.value) {
+                        { Icon(imageVector = Icons.Filled.Done, contentDescription = "Done icon") }
+                    } else null
+                )
+
+                FilterChip(
+                    onClick = { favoriteState.value = !favoriteState.value },
+                    label = { Text("Add to Favorites") },
+                    selected = favoriteState.value,
+                    leadingIcon = if (favoriteState.value) {
+                        { Icon(imageVector = Icons.Filled.Done, contentDescription = "Done icon") }
+                    } else null
+                )
+            }
+
         }
-        Row(
-            modifier = modifier.padding(5.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            Checkbox(
-                modifier = Modifier.padding(5.dp),
-                checked = favoriteState.value,
-                onCheckedChange = {
-                    favoriteState.value = it
-                    password.isFavorite = it
-                }
-            )
-            Text(
-                text = "Add to Favorites",
-                modifier = Modifier.padding(5.dp)
-            )
-            Text(text = "Lock this password with Biometric",modifier = Modifier.padding(5.dp))
-            Switch(
-                modifier = Modifier.padding(5.dp),
-                checked = enableBiometricAuthState.value,
-                onCheckedChange = { enableBiometricAuthState.value = it }
-            )
-        }
-        Row(modifier = modifier.fillMaxWidth().padding(start = 20.dp,end = 10.dp)){
-            Button(
+        Row(modifier = modifier.padding(start = 10.dp)) {
+            FilledTonalButton(
                 onClick = {
-                    if(validatePassword && validateUserName && validateTypeState) {
+                    if (validatePassword && validateUserName && validateTypeState) {
                         password.username = userNameState.value
                         password.password = CryptoUtils().encrypt(passwordState.value)
                         password.description = descriptionState.value
@@ -155,11 +163,10 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
                         password.enableBiometricAuth = enableBiometricAuthState.value
                         password.updatedAt = Date()
                         passwordViewModel.updatePassword(password)
-                        Toast.makeText(context, "Password Edited Successfully", Toast.LENGTH_SHORT)
+                        Toast.makeText(context, "Password Updated Successfully", Toast.LENGTH_SHORT)
                             .show()
                         navController.popBackStack()
-                    }
-                    else{
+                    } else {
                         Toast.makeText(context, "Please fill all the fields", Toast.LENGTH_SHORT)
                             .show()
                     }
@@ -167,18 +174,27 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
                 },
                 modifier = modifier.padding(5.dp)
             ) {
-                Icon(imageVector = Icons.Filled.Update, contentDescription = "Update Password",modifier = modifier.size(
-                    ButtonDefaults.IconSize))
+                Icon(
+                    imageVector = Icons.Filled.Update,
+                    contentDescription = "Update Password",
+                    modifier = modifier.size(
+                        ButtonDefaults.IconSize
+                    )
+                )
                 Spacer(modifier = modifier.size(ButtonDefaults.IconSpacing))
                 Text(text = "Update")
 
             }
-            Button(
+            FilledTonalButton(
                 onClick = {
-                    passwordViewModel.deletePassword(password)
-                    Toast.makeText(context,"Password Deleted Successfully",Toast.LENGTH_SHORT).show()
-                    navController.popBackStack()
-
+                    if(password.enableBiometricAuth){
+                        showBiometricAuthDialog.value = true
+                    }
+                    else{
+                        passwordViewModel.deletePassword(password)
+                        Toast.makeText(context, "Password Deleted Successfully", Toast.LENGTH_SHORT).show()
+                        navController.popBackStack()
+                    }
                 },
                 modifier = modifier.padding(5.dp),
                 colors = ButtonColors(
@@ -187,13 +203,36 @@ fun EditPasswordFormComponent(modifier: Modifier = Modifier, password: Password,
                     disabledContentColor = Color.White.copy(alpha = 0.1f)
                 )
             ) {
-                Icon(imageVector = Icons.Filled.Delete, contentDescription = "Delete Password",modifier = modifier.size(
-                    ButtonDefaults.IconSize))
+                Icon(
+                    imageVector = Icons.Filled.Delete,
+                    contentDescription = "Delete Password",
+                    modifier = modifier.size(
+                        ButtonDefaults.IconSize
+                    )
+                )
                 Spacer(modifier = modifier.size(ButtonDefaults.IconSpacing))
                 Text(text = "Delete")
 
             }
         }
+    }
+    if (showBiometricAuthDialog.value) {
+        BiometricAuthComponent(
+            onSuccess = {
+
+                passwordViewModel.deletePassword(password)
+                Toast.makeText(context, "Password Deleted Successfully", Toast.LENGTH_SHORT).show()
+                navController.popBackStack()
+                showBiometricAuthDialog.value = false
+            },
+            onError = {
+
+                Toast.makeText(context, "Unable to Delete Password: Authentication Failed", Toast.LENGTH_SHORT).show()
+                navController.navigate(Routes.PasswordDetailScreen.name + "/${password.id}")
+                showBiometricAuthDialog.value = false
+            },
+            subtitle = "Authenticate to Delete the password"
+        )
     }
 
 }
